modal for customer view added, functionality not working yet.
For functionality I was thinking about printing out a table with just
customer first and last name with 2 buttons. One for viewing specific
details about the customer and another for editing the customer's info.

They're both going to be bootstrap modals with AJAX calls to pull and
edit the right information.